Implement aes_ctr method for iterators
This was actually a lot harder than I thought it would be, and I ended
up escaping with a Box<dyn Iterator> to mask the opaque type returned by
aes::ctr. If I were to do this without a Box, I would have to store the
upstream iterator and all the state associated with encrypting block by
block and emitting bytes. I wonder if I could use Flatten for that?